Added changes w.r.t proposal and moved batching into seperate class
